parsers/site2_parser.py
from bs4 import BeautifulSoup

def parse_books(html):
    soup = BeautifulSoup(html, "html.parser")
    data = []

    for book in soup.select(".product_pod"):
        title = book.h3.a["title"]
        price = book.select_one(".price_color").get_text()

        data.append({
            "type": "book",
            "title": title,
            "price": price
        })

    return data