requirements.txt
requests
ddgs
beautifulsoup4
pandas
openpyxl
fpdf2
pymongo